+++ flamingo/lib/src/model/collection.dart
@@ -4,11 +4,11 @@ import '../type/type.dart';
 import 'document.dart';
 
 class Collection<T extends Document<DocumentType>> {
-  Collection(Document parent, this.name) {
-    ref = parent.reference.collection(name);
-    path = ref.path;
-  }
-  CollectionReference ref;
-  String path;
-  String name;
+  Collection(Document parent, this.name)
+      : ref = parent.reference.collection(name),
+        path = parent.reference.collection(name).path;
+
+  final CollectionReference ref;
+  final String path;
+  final String name;
 }