+++ flamingo/lib/src/model/storage_file.dart
@@ -1,26 +1,29 @@
+import 'package:flutter/foundation.dart';
+
 part 'storage_file.g.dart';
 
 class StorageFile {
   StorageFile(
-      {this.name,
-      this.path,
-      this.url,
+      {@required this.name,
+      @required this.path,
+      @required this.url,
       this.mimeType = 'application/octet-stream',
       this.additionalData = const <String, dynamic>{},
       this.metadata = const <String, String>{}});
   factory StorageFile.fromJson(Map<String, dynamic> json) =>
       _$StorageFileFromJson(json);
 
-  bool isDeleted = false;
+  bool get isDeleted => _isDeleted;
+  bool _isDeleted = false;
 
-  String folderName;
-
-  String name;
-  String path;
-  String url;
-  String mimeType;
-  Map<String, dynamic> additionalData;
-  Map<String, String> metadata;
+  final String name;
+  final String path;
+  final String url;
+  final String mimeType;
+  final Map<String, dynamic> additionalData;
+  final Map<String, String> metadata;
 
   Map<String, dynamic> toJson() => _$StorageFileToJson(this);
+
+  void deleted() => _isDeleted = true;
 }